# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Naam Jaap Sewa Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sankalp/today:
    get:
      tags: [Sankalp]
      operationId: getTodaySankalp
      summary: Get today's sankalp status for the user
      responses:
        '200':
          description: Today's sankalp envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SankalpEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Sankalp]
      operationId: acceptTodaySankalp
      summary: Accept today's sankalp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptSankalpBody'
      responses:
        '200':
          description: Sankalp accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SankalpEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /sankalp/history:
    get:
      tags: [Sankalp]
      operationId: getSankalpHistory
      summary: Recent sankalp history (last 30 days)
      responses:
        '200':
          description: Sankalp history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SankalpHistoryEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Sankalp:
      type: object
      required: [id, userId, date, accepted, acceptedAt, text]
      properties:
        id:
          type: string
        userId:
          type: string
        date:
          type: string
        accepted:
          type: boolean
        acceptedAt:
          type: [string, 'null']
          format: date-time
        text:
          type: string
    SankalpEnvelope:
      type: object
      required: [sankalp, alreadyAcceptedToday]
      properties:
        sankalp:
          $ref: '#/components/schemas/Sankalp'
        alreadyAcceptedToday:
          type: boolean
    SankalpHistoryItem:
      type: object
      required: [date, accepted]
      properties:
        date:
          type: string
        accepted:
          type: boolean
        acceptedAt:
          type: [string, 'null']
          format: date-time
    SankalpHistoryEnvelope:
      type: object
      required: [history]
      properties:
        history:
          type: array
          items:
            $ref: '#/components/schemas/SankalpHistoryItem'
    AcceptSankalpBody:
      type: object
      properties:
        text:
          type: [string, 'null']
...